all_the_buzz/requirements.txt
requests==2.32.5
pyaml==25.7.0
PyYAML==6.0.3
Sphinx==8.2.3
nh3==0.3.1
pymongo==4.15.3
flask==3.1.2
pylint==4.0.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.2.1
colorama==0.4.6
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
style==1.1.0
update==0.0.1
validators==0.35.0
sphinx_rtd_theme
